(#3) Use replace instead of add/delete if possible
In a modification where none of the current values are to be retained rather
than delete each value and then add the new value(s), use a replace operation
with the new value(s) which will implicitly remove any old value.

This means replace will always be used when replacing a single-valued
attribute and should mean attributes such as olcTLSCertificateFile can be
changed.