(#15) Fix idempotency for olcSyncrepl attribute
The positional `{x}` prefix was not being explicitly added to the attribute
values but slapd was adding them resulting in a difference on each run.

The test case was wrong (now corrected) and the `openldap_values()` function
has been updated to prune out any nil or empty string values prior to adding
the prefix.